更新

+++ internal/model/entity/cms_role.go
@@ -6,7 +6,7 @@ package entity
 
 // CmsRole is the golang structure for table cms_role.
 type CmsRole struct {
-	Id          uint   `json:"id"          ` //
+	Id          uint64 `json:"id"          ` //
 	Name        string `json:"name"        ` // 角色组名称
 	Title       string `json:"title"       ` // 中文名称
 	Description string `json:"description" ` //
